+++ api/src/main/java/uk/gov/hmcts/payment/api/service/IACServiceImpl.java
@@ -1,25 +1,15 @@
 package uk.gov.hmcts.payment.api.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixProperty;
 import com.netflix.hystrix.exception.HystrixRuntimeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestClientException;
-import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.payment.api.contract.PaymentDto;
 import uk.gov.hmcts.payment.api.dto.*;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -27,18 +17,8 @@ import java.util.stream.Collectors;
 public class IACServiceImpl implements IACService {
     private static final Logger LOG = LoggerFactory.getLogger(IACServiceImpl.class);
 
-    @Value("${iac.supplementary.info.url}")
-    private String iacSupplementaryInfoUrl;
-
     @Autowired
-    @Qualifier("restTemplateIacSupplementaryInfo")
-    private RestTemplate restTemplateIacSupplementaryInfo;
-
-    @Autowired
-    private AuthTokenGenerator authTokenGenerator;
-
-    @Autowired
-    private IACService iacService;
+    private IACServiceResponse iacServiceResponse;
 
     @Override
     public ResponseEntity<SupplementaryPaymentDto> getIacSupplementaryInfo(List<PaymentDto> paymentDtos, String serviceName) {
@@ -50,6 +30,7 @@ public class IACServiceImpl implements IACService {
 
         List<String> iacCcdCaseNos = iacPayments.stream().map(paymentDto -> paymentDto.getCcdCaseNumber()).
             collect(Collectors.toList());
+
         ResponseEntity<SupplementaryDetailsResponse> responseEntitySupplementaryInfo = null;
 
         List<SupplementaryInfo> lstSupplementaryInfo = null;
@@ -58,7 +39,7 @@ public class IACServiceImpl implements IACService {
         if (!iacCcdCaseNos.isEmpty()) {
             LOG.info("List of IAC Ccd Case numbers : {}", iacCcdCaseNos.toString());
             try {
-                responseEntitySupplementaryInfo = iacService.getIacSupplementaryInfoResponse(iacCcdCaseNos);
+                responseEntitySupplementaryInfo = iacServiceResponse.getIacSupplementaryInfoResponse(iacCcdCaseNos);
             } catch (HttpClientErrorException ex) {
                 LOG.info("IAC Supplementary information could not be found, exception: {}", ex.getMessage());
                 paymentResponseHttpStatus = HttpStatus.PARTIAL_CONTENT;
@@ -93,31 +74,6 @@ public class IACServiceImpl implements IACService {
                     build();
             }
            return new ResponseEntity(supplementaryPaymentDto, paymentResponseHttpStatus);
-    }
-
-    @Override
-    @HystrixCommand(commandKey = "IACService", commandProperties = {
-        @HystrixProperty(name = "execution.isolation.thread.timeoutInMilliseconds", value = "10000"),
-        @HystrixProperty(name = "fallback.enabled", value = "false")
-    })
-     public ResponseEntity<SupplementaryDetailsResponse> getIacSupplementaryInfoResponse(List<String> iacCcdCaseNos) throws RestClientException {
-
-        IacSupplementaryRequest iacSupplementaryRequest = IacSupplementaryRequest.createIacSupplementaryRequestWith()
-            .ccdCaseNumbers(iacCcdCaseNos).build();
-
-        MultiValueMap<String, String> headerMultiValueMapForIacSuppInfo = new LinkedMultiValueMap<String, String>();
-        List<String> serviceAuthTokenPaymentList = new ArrayList<>();
-
-         //Generate token for payment api and replace
-         serviceAuthTokenPaymentList.add(authTokenGenerator.generate());
-         LOG.info("S2S Token To Test : {}", authTokenGenerator.generate());
-
-        headerMultiValueMapForIacSuppInfo.put("ServiceAuthorization", serviceAuthTokenPaymentList);
-        LOG.info("IAC Supplementary info URL: {}", iacSupplementaryInfoUrl + "/supplementary-details");
-
-        HttpHeaders headers = new HttpHeaders(headerMultiValueMapForIacSuppInfo);
-        final HttpEntity<IacSupplementaryRequest> entity = new HttpEntity<>(iacSupplementaryRequest, headers);
-        return this.restTemplateIacSupplementaryInfo.exchange(iacSupplementaryInfoUrl + "/supplementary-details", HttpMethod.POST, entity, SupplementaryDetailsResponse.class);
     }
 
 }
