+++ api/src/main/java/uk/gov/hmcts/payment/api/service/IACServiceResponse.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.payment.api.service;
+import org.springframework.http.ResponseEntity;
+import uk.gov.hmcts.payment.api.dto.SupplementaryDetailsResponse;
+import java.util.List;
+
+public interface IACServiceResponse {
+    ResponseEntity<SupplementaryDetailsResponse> getIacSupplementaryInfoResponse(List<String> iacCcdCaseNos);
+}
