+++ cve-resolution-strategy.gradle
@@ -36,13 +36,13 @@ configurations.all {
             }
 
             /*
-                CVE-2021-28165, CVE-2021-28163, CVE-2021-28164
+                CVE-2021-28165, CVE-2021-28163, CVE-2021-28164, CVE-2021-28165
             * */
-            if(det.requested.name == 'jetty-client' || det.requested.name == 'jetty-continuation' || det.requested.name == 'jetty-http'
-                    ||det.requested.name == 'jetty-security' ||det.requested.name == 'jetty-servlet' ||det.requested.name == 'jetty-servlets'
-                    ||det.requested.name == 'jetty-util' ||det.requested.name == 'jetty-webapp' ||det.requested.name == 'jetty-xml'
-                    ||det.requested.name == 'jetty-server' ||det.requested.name == 'jetty-_http_server' ||det.requested.name == 'jetty-util-ajax'
-                    ||det.requested.name == 'mortbay_jetty'){
+            if (det.requested.name == 'jetty-client' || det.requested.name == 'jetty-continuation' || det.requested.name == 'jetty-http'
+                    || det.requested.name == 'jetty-security' || det.requested.name == 'jetty-servlet' || det.requested.name == 'jetty-servlets'
+                    || det.requested.name == 'jetty-util' || det.requested.name == 'jetty-webapp' || det.requested.name == 'jetty-xml'
+                    || det.requested.name == 'jetty-server' || det.requested.name == 'jetty-_http_server' || det.requested.name == 'jetty-util-ajax'
+                    || det.requested.name == 'mortbay_jetty' || det.requested.name == 'jetty-io') {
                 det.useVersion '9.4.39.v20210325'
             }
 
@@ -50,8 +50,8 @@ configurations.all {
                 CVE-2021-22112
             **/
 
-            if(det.requested.name == 'spring-security-config' || det.requested.name == 'spring-security-core' || det.requested.name == 'spring-security-crypto'
-                    || det.requested.name == 'spring-security-web'){
+            if (det.requested.name == 'spring-security-config' || det.requested.name == 'spring-security-core' || det.requested.name == 'spring-security-crypto'
+                    || det.requested.name == 'spring-security-web') {
                 det.useVersion '5.4.5'
             }
 
@@ -107,21 +107,21 @@ configurations.all {
             /*
                 CVE-2017-18640
             * */
-            if(det.requested.name == 'snakeyaml'){
+            if (det.requested.name == 'snakeyaml') {
                 det.useVersion '1.27'
             }
 
             /*
                 CVE-2012-6708, CVE-2015-9251, CVE-2019-11358
             **/
-            if(det.requested.name == 'ff4j-web'){
+            if (det.requested.name == 'ff4j-web') {
                 det.useVersion '1.8.11'
             }
 
             /*
                 CWE-79: Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')
             * */
-            if(det.requested.name == 'hystrix-core'){
+            if (det.requested.name == 'hystrix-core') {
                 det.useVersion '1.5.18'
             }
